Refactor functions so they are not polluting the render code
